+++ index-comedor/comedor.css
@@ -17,7 +17,13 @@ body{
     font-weight:800;
     font-family: "Lucida Console", Courier, monospace;
     margin-left: 2vw;
+}
 
+.container-row{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
 }
 
 .logo-Unco{
@@ -60,7 +66,7 @@ body{
     margin: 2%
 }
 
-.container-row{
+.container-row-food{
     display: flex;
     flex-direction: row;
     align-items: center;
@@ -94,6 +100,14 @@ body{
     color: white;
     font-weight: 500;
 }
+
+#clickeable:hover{
+
+}
+
+/*
+
+*/
 
 .little-box{
     position: relative;
@@ -175,16 +189,20 @@ body{
     width: 900px;
     height: 680px;
     font-size: 0.40cm;
-    background-color:white;
-    border:4px solid rgb(87, 86, 86);
+    background: rgba(179, 179, 179, 0.2);
+    border-radius: 16px;
+    box-shadow: 10px 4px 30px rgba(0, 0, 0, 0.1);
+    backdrop-filter: blur(50px);
+    border: 1px solid rgba(255, 255, 255, 0.3);
 }
+
 .big-menu-title{
     font-size: 3em;
     font-family: 'Bevan';
 }
 
 .large-menu-paragraph{
-    font-size: 1.2em;
+    font-size: 1.5em;
 }
 
 .price{
@@ -194,6 +212,7 @@ body{
     font-weight: 800;
     text-align: right;
     margin-right: 20px;
+    backdrop-filter: blur(0px);
 }
 
 #button-to-reserve {
@@ -222,15 +241,108 @@ body{
     flex-direction: row;
 }
 
-#like-button{
-    display: block;
-    font-size: 52px; 
-    border: solid; 
-    padding: 10px;
-    margin: 10px; 
-    border-radius: 50%;
-     background-color: rgb(255, 255, 255); 
-     color: rgb(0, 0, 0);
+/*
+#like-button {
+  all: unset;
+  width: 10px;
+  height: 30px;
+  font-size: 16px;
+  background: transparent;
+  border: none;
+  position: relative;
+  color: #f0f0f0;
+  cursor: pointer;
+  z-index: 1;
+  padding: 10px 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  white-space: nowrap;
+}
+
+#like-button::after,
+#like-button::before {
+  content: '';
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  z-index: -99999;
+  transition: all .4s;
+}
+
+#like-button::before {
+  transform: translate(0%, 0%);
+  width: 100%;
+  height: 100%;
+  background: #28282d;
+  border-radius: 10px;
+}
+
+#like-button::after {
+  transform: translate(10px, 10px);
+  width: 35px;
+  height: 35px;
+  background: #ffffff15;
+  backdrop-filter: blur(5px);
+  border-radius: 50px;
+  border-radius: 10px;
+  transform: translate(0, 0);
+  width: 100%;
+  height: 100%;
+}
+
+#like-button:hover::before {
+  transform: translate(5%, 20%);
+  width: 110%;
+  height: 110%;
+}
+
+
+#like-button:active::after {
+  transition: 0s;
+  transform: translate(0, 5%);
+}
+*/
+#like-button {
+  align-items: center;
+  appearance: none;
+  background-image: radial-gradient(100% 100% at 100% 0, #5adaff 0, #5468ff 100%);
+  border: 0;
+  border-radius: 6px;
+  box-shadow: rgba(45, 35, 66, .4) 0 2px 4px,rgba(45, 35, 66, .3) 0 7px 13px -3px,rgba(58, 65, 111, .5) 0 -3px 0 inset;
+  box-sizing: border-box;
+  color: #fff;
+  cursor: pointer;
+  display: inline-flex;
+  height: 48px;
+  justify-content: center;
+  line-height: 1;
+  list-style: none;
+  overflow: hidden;
+  padding-left: 16px;
+  padding-right: 16px;
+  position: relative;
+  text-align: left;
+  text-decoration: none;
+  transition: box-shadow .15s,transform .15s;
+  touch-action: manipulation;
+  white-space: nowrap;
+  will-change: box-shadow,transform;
+  font-size: 18px;
+}
+
+#like-button:focus {
+  box-shadow: #3c4fe0 0 0 0 1.5px inset, rgba(45, 35, 66, .4) 0 2px 4px, rgba(45, 35, 66, .3) 0 7px 13px -3px, #3c4fe0 0 -3px 0 inset;
+}
+
+#like-button:hover {
+  box-shadow: rgba(45, 35, 66, .4) 0 4px 8px, rgba(45, 35, 66, .3) 0 7px 13px -3px, #3c4fe0 0 -3px 0 inset;
+  transform: translateY(-2px);
+}
+
+#like-button:active {
+  box-shadow: #3c4fe0 0 3px 7px inset;
+  transform: translateY(2px);
 }
 
 #close{
