--- index/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Nombre de la pagina</title>
-        <meta charset="utf-8">
-        <meta lang="es">
-        <link rel="stylesheet" href="index.css">
-        <!--Los siguientes links son de bootstrap-->
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
-    </head>
-    <body>
-        <div class="container-menus">
-            <h3 id="time-foods">Desayuno - Horarios 9:30-11</h3>
-            <section id="container-row-breakfast">
-                
-            </section>
-            <h3 id="time-foods">Almuerzo - Horarios 12-14</h3>
-            <section id="container-row-lunch">
-
-            </section>
-            <h3 id="time-foods">Merienda - Horarios 17-19</h4>
-            <section id="container-row-merienda"></section>
-            <script type="module" src="index.js"></script>
-          </div>
-    </body>
-</html>
