+++ proxComedor/proxComidas.css
@@ -1,0 +1,248 @@
+*{
+    margin: 0;
+    padding: 0;
+}
+
+body{
+    margin: auto;
+    background-color: rgb(41, 40, 40);
+    display: grid;
+    min-height: 100vh;
+    grid-template-rows: auto 1fr auto;
+}
+
+/*Encabezado*/
+#title-Unco{
+    font-size: 2vw;
+    font-weight:800;
+    font-family: "Lucida Console", Courier, monospace;
+    margin-left: 2vw;
+
+}
+
+.logo-Unco{
+    width: 6vw;
+    height:6vw;
+    margin-left: 1vw;
+    margin-right: 12vw;
+}
+
+/*Imagen de fondo del menu*/
+.background-img-menu{
+    background-image: url("https://10619-2.s.cdn12.com/rests/original/104_355872023.jpg");
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-position: center;
+}
+
+#darken-background{
+    width : 100%;
+	height : 100%;
+    background-color: rgb(41, 40, 40);
+    opacity: 0.9;
+}
+
+/*Contenedor del titulo de comida hoy*/
+.title-menu-today{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 100px;
+    max-width: 100%;
+    color: white;
+    font-weight: 600;
+    font-size: 3.5vw;
+
+}
+
+/*Contenedor de las comidas*/
+.container-menus{
+    margin: 2%
+}
+
+.container-row{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+}
+
+/*Eslogan*/
+.tip{
+    display: flex;
+    justify-content: center;
+    color: white;
+    font-weight: 600;
+    padding: 2vw;
+}
+
+.people-eat{
+    filter: brightness(0%);
+    width: 20vw;
+    height: 20vw;
+}
+
+.fork{
+    width: 20vw;
+    height: 20vw;
+}
+
+/*Menu de desayuno,almuerzo,merienda*/
+#time-foods{
+    display: flex;
+    justify-content: center;
+    color: white;
+    font-weight: 500;
+}
+
+.little-box{
+    position: relative;
+    display: inline-block;
+    text-align: center;
+    max-height: 40vh;
+    max-width: 40vw;
+    overflow: hidden;
+}
+
+.little-box-title{
+    max-height: 100%;
+    max-width: 100%;
+    position: absolute;
+    top: 80%;
+    left: 50%;
+    transform: translate(-80%, -50%);
+}
+
+#little-title-style{
+    color: antiquewhite;
+    text-shadow: 4px 4px 2px black;
+}
+
+.small-image{
+    max-height: 100%;
+    max-width: 100%;
+}
+
+@media (max-width: 6px){
+    #little-title-style{
+        font-size: 0.5vw;
+    }
+
+    .little-box-title{
+        max-height: 100%;
+        max-width: 100%;
+        position: absolute;
+        top: 70%;
+        left: 50%;
+        transform: translate(-70%, -50%);
+    }
+}
+
+/*Proximas comidas*/
+.next-meals{
+    align-items: center;
+    display: flex;
+    gap: 0.25rem;
+    justify-content: center;
+    position: relative;
+    font-size: 2vw;
+}
+
+/*Datos de contacto*/
+.contact{
+    background-color: #101214;
+    height: 150px;
+    max-width: 100%;
+}
+
+/*ESTILO DEL MENU AMPLIADO*/
+
+.big-image{
+    width: 892px;
+    height: 300px;
+    margin: auto;
+    filter: brightness(0.5);
+}
+
+.large-menu-box{
+    position: absolute;
+    z-index: 1;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    margin: auto;
+    width: 900px;
+    height: 680px;
+    font-size: 0.40cm;
+    background-color:white;
+    border:4px solid rgb(87, 86, 86);
+}
+.big-menu-title{
+    font-size: 3em;
+    font-family: 'Bevan';
+}
+
+.large-menu-paragraph{
+    font-size: 1.2em;
+}
+
+.price{
+    color: rgb(4, 123, 197);
+    font-size: 1.7em;
+    font-family: Vegur, 'PT Sans', Verdana, sans-serif;
+    font-weight: 800;
+    text-align: right;
+    margin-right: 20px;
+}
+
+#button-to-reserve {
+    width: 280px;
+    height: 80px;
+    margin: 10px;
+    background: rgb(4, 123, 197);
+    border-radius: 6px;
+    font-size: 2em;
+    font-family: Vegur, 'PT Sans', Verdana, sans-serif;
+    color: white;
+    border-radius: 15px;
+    float: right;
+  }
+
+#button-to-reserve:hover {background-color:rgb(2, 67, 107)}
+
+#button-to-reserve:active {
+  background-color: rgb(4, 123, 197);
+  box-shadow: 0 5px #666;
+  transform: translateY(4px);
+}
+
+.conteiner-row{
+    display: flex;
+    flex-direction: row;
+}
+
+#like-button{
+    display: block;
+    font-size: 52px; 
+    border: solid; 
+    padding: 10px;
+    margin: 10px; 
+    border-radius: 50%;
+     background-color: rgb(255, 255, 255); 
+     color: rgb(0, 0, 0);
+}
+
+#close{
+    position: absolute; 
+    right: 20px; top: 13px; width:20px; 
+    background-color: #FFF; 
+    color: black; 
+    margin-top:-15px; margin-right:-15px; border-radius: 20px; 
+    padding-left: 3px; padding-top: 1px; 
+    cursor:pointer; 
+    z-index: 2; 
+    font-size:16px; 
+    font-weight:bold;
+
+}
